backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0
httpx>=0.27.0
numpy>=1.26.0
Pillow>=10.0.0
ruff>=0.3.0
mypy>=1.8.0
